Fixed uninitialized variable
I experienced graphics problems requiring hard restarts after opening the Stars scene followed by the voxel scene.
The reason was that the number of faces was not initalized on the first run, it would only be initialized after the first render.
Now the value gets initalized and the faces generated before rendering.